Update Tasks / ClusterTasks views to be consistent with PipelineRuns
- Use labelSelector to filter
- Use list view when displaying multiple runs
- TaskRun view only used to view a single run
- Rename TaskRunList container to TaskRuns for consistency with PipelineRuns
- Remove duplicate request for PipelineRuns when removing a filter
