+++ package.json
@@ -33,6 +33,7 @@
     "history": "^4.9.0",
     "isomorphic-fetch": "^2.2.1",
     "js-yaml": "^3.13.0",
+    "lodash.isequal": "^4.5.0",
     "lodash.keyby": "^4.6.0",
     "lodash.merge": "^4.6.2",
     "lodash.snakecase": "^4.1.1",
