+++ src/api/index.js
@@ -114,20 +114,10 @@ export function getPipeline({ name, namespace }) {
   return get(uri);
 }
 
-export function getPipelineRuns({
-  namespace,
-  pipelineName,
-  filters = []
-} = {}) {
+export function getPipelineRuns({ filters = [], namespace } = {}) {
   let queryParams;
-  if (filters.length > 0 || pipelineName) {
-    queryParams = { labelSelector: [] };
-    if (pipelineName) {
-      queryParams.labelSelector.push(`tekton.dev/pipeline=${pipelineName}`);
-    }
-    if (filters.length > 0) {
-      queryParams.labelSelector.push(filters);
-    }
+  if (filters.length) {
+    queryParams = { labelSelector: filters };
   }
   const uri = getTektonAPI('pipelineruns', { namespace }, queryParams);
   return get(uri).then(checkData);
@@ -217,11 +207,12 @@ export function getTask({ name, namespace }) {
   return get(uri);
 }
 
-export function getTaskRuns({ namespace, taskName } = {}) {
+export function getTaskRuns({ filters = [], namespace } = {}) {
   let queryParams;
-  if (taskName) {
-    queryParams = { labelSelector: `tekton.dev/task=${taskName}` };
+  if (filters.length) {
+    queryParams = { labelSelector: filters };
   }
+
   const uri = getTektonAPI('taskruns', { namespace }, queryParams);
   return get(uri).then(checkData);
 }
