+++ src/store/middleware.js
@@ -11,6 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+/* istanbul ignore next */
 export function createWebSocketMiddleware(socket) {
   return ({ dispatch }) => {
     socket.addEventListener('close', () => {
